Move label map to avoid cyclic import
